+++ packages/orq-rc/src/models/operations/getbudget.ts
@@ -296,7 +296,7 @@ export const GetBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-12-10T05:40:55.012Z",
+    "2025-12-10T08:38:35.548Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -336,7 +336,7 @@ export const GetBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => GetBudgetConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-12-10T05:40:55.012Z"))
+  updated: z.date().default(() => new Date("2025-12-10T08:38:35.548Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {