+++ packages/orq-rc/src/models/operations/updatebudget.ts
@@ -441,7 +441,7 @@ export const UpdateBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-12-10T05:40:55.012Z",
+    "2025-12-10T08:38:35.548Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -481,7 +481,7 @@ export const UpdateBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => UpdateBudgetConsumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-12-10T05:40:55.012Z"))
+  updated: z.date().default(() => new Date("2025-12-10T08:38:35.548Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {