+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -335,7 +335,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kc3ce28z7j1rxx1axnrez39h"),
+  _id: z.string().default("conv_01kc3pkd4chwb8emxha7g9cr7k"),
   entity_id: z.string(),
   kind: UpdateConversationKind$inboundSchema,
   display_name: z.string(),
@@ -375,7 +375,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kc3ce28z7j1rxx1axnrez39h"),
+  id: z.string().default("conv_01kc3pkd4chwb8emxha7g9cr7k"),
   entityId: z.string(),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),